+++ src/modules/common/components/header.component.tsx
@@ -8,8 +8,8 @@ import {
     Container,
     Drawer,
     IconButton,
-    Tab as MuiTab,
     Stack,
+    Tab as MuiTab,
     Tabs,
     Toolbar,
     Typography,
