+++ src/modules/route-services/components/index.ts
@@ -1,8 +1,8 @@
 export * from './nav.component';
 export * from './search-result.component';
-export * from './service-create.component';
 export * from './service-listing.component';
-export * from './service-listings.component';
 export * from './service-notification.component';
-export * from './service-notifications.component';
 export * from './service-tags.component';
+export * from './service-notifications.component'
+export * from './service-create.component'
+export * from './service-listings.component'
