+++ src/modules/route-services/components/service-tags.component.tsx
@@ -22,19 +22,19 @@ export const ServiceTags: React.FC = () => {
 
     const { i18n } = useLingui();
 
-        return (
-            <ToggleButtonGroup
-                value={selectedTags}
-                onChange={handleTagSelection}
-                aria-label={t(i18n)`Service Types`}
-            >
-                <ToggleButton value={TAGS.POWERLEVELING}>{t(i18n)`Powerleveling`}</ToggleButton>
-                <ToggleButton value={TAGS.BOSS_HELP}>{t(i18n)`Boss Help`}</ToggleButton>
-                <ToggleButton value={TAGS.UBER_LILITH}>{t(i18n)`Uber Lilith`}</ToggleButton>
-                <ToggleButton value={TAGS.CAPSTONE_BOOST}>{t(i18n)`Capstone Boost`}</ToggleButton>
-                <ToggleButton value={TAGS.EUROPE}>{t(i18n)`Europe`}</ToggleButton>
-                <ToggleButton value={TAGS.ASIA}>{t(i18n)`Asia`}</ToggleButton>
-                <ToggleButton value={TAGS.AMERICA}>{t(i18n)`America`}</ToggleButton>
-            </ToggleButtonGroup>
-        );
+    return (
+        <ToggleButtonGroup
+            value={selectedTags}
+            onChange={handleTagSelection}
+            aria-label={t(i18n)`Service Types`}
+        >
+            <ToggleButton value={TAGS.POWERLEVELING}>{t(i18n)`Powerleveling`}</ToggleButton>
+            <ToggleButton value={TAGS.BOSS_HELP}>{t(i18n)`Boss Help`}</ToggleButton>
+            <ToggleButton value={TAGS.UBER_LILITH}>{t(i18n)`Uber Lilith`}</ToggleButton>
+            <ToggleButton value={TAGS.CAPSTONE_BOOST}>{t(i18n)`Capstone Boost`}</ToggleButton>
+            <ToggleButton value={TAGS.EUROPE}>{t(i18n)`Europe`}</ToggleButton>
+            <ToggleButton value={TAGS.ASIA}>{t(i18n)`Asia`}</ToggleButton>
+            <ToggleButton value={TAGS.AMERICA}>{t(i18n)`America`}</ToggleButton>
+        </ToggleButtonGroup>
+    );
 };
